requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
joblib==1.3.2
urllib3==2.0.4
lxml==4.9.3
html5lib==1.1
python-pptx==0.6.21
